re-organized the Transform folder in preparation for major refactor to fix the Experience types
