+++ backend/src/extract-profile/transform/profile/getJobDuration.ts
