+++ backend/src/extract-profile/transform/profile/verifyImage.test.ts
