+++ backend/src/extract-profile/transform/profile/verifyImage.ts
