+++ backend/src/extract-profile/transform/report/checkEmptyFields.ts
