+++ backend/src/extract-profile/transform/report/checkPrivateProfile.test.ts
@@ -3,10 +3,10 @@ import { expect, test } from "vitest";
 
 // Project files
 import checkPrivateProfile from "./checkPrivateProfile";
-import ReportLog from "../../types/ReportLog";
-import ReportSeverity from "../../types/ReportSeverity";
-import { privateProfile } from "../_test-websites/profilePrivate";
-import { profile1 } from "../_test-websites/profile1";
+import ReportLog from "../../../types/ReportLog";
+import ReportSeverity from "../../../types/ReportSeverity";
+import { privateProfile } from "../../_test-websites/profilePrivate";
+import { profile1 } from "../../_test-websites/profile1";
 
 test("Returns the correct type of report when sent a private LinkedIn page", () => {
   // Arrange
